Added additional quality copts (#2)
* Re-enabling werror and enabling wall, wextra, and pendantic

* Fixing warnings (treated as errors) that were only visible on linux

* Removing debug code used in troubleshooting

* Adding openvx integration test in CI

* Fixed issue with linking on linux

+++ kernels/utils/conversion_utils.cpp
@@ -136,7 +136,7 @@ int16_t conversion_24_8(int32_t acc_value)
     }
     else
     {
-        out = 0x8000;
+        out = static_cast<int16_t>(0x8000);
     }
     return(out);
 }
